Rename branching_scheme* files to tree_search* across all domains
Renames branching_scheme.{hpp,cpp} to tree_search.{hpp,cpp} in all six
domains (onedimensional, rectangle, rectangleguillotine, box, boxstacks,
irregular), and branching_scheme_maximal_spaces.{hpp,cpp} to
tree_search_maximal_spaces.{hpp,cpp} in box, rectangle, and
rectangleguillotine. Updates CMakeLists.txt source lists and all
#include directives in optimize.cpp, test files, and internal headers.
No functional changes.
