+++ src/boxstacks/CMakeLists.txt
@@ -9,7 +9,7 @@ target_sources(PackingSolver_boxstacks PRIVATE
     optimize.cpp
     post_process.cpp
     sequential_onedimensional_rectangle.cpp
-    branching_scheme.cpp)
+    tree_search.cpp)
 target_include_directories(PackingSolver_boxstacks PUBLIC
     ${PROJECT_SOURCE_DIR}/include)
 target_include_directories(PackingSolver_boxstacks PRIVATE
