+++ src/rectangle/tree_search_maximal_spaces.hpp
