+++ src/rectangleguillotine/tree_search.hpp
