+++ src/rectangleguillotine/tree_search_maximal_spaces.hpp
