+++ test/rectangleguillotine/branching_scheme/trim_test.cpp
@@ -1,5 +1,5 @@
 #include "packingsolver/rectangleguillotine/instance_builder.hpp"
-#include "rectangleguillotine/branching_scheme.hpp"
+#include "rectangleguillotine/tree_search.hpp"
 
 #include <gtest/gtest.h>
